+++ src/components/Help.jsx
@@ -39,48 +39,107 @@ const Help = () => {
     <div class="help">
       <h1>Help!</h1>
       <p>
-        PIKSEQ is a mix between a pixel art editor and a sequencer.
+        PIKSEQ is a <del>stupid</del> fun mix between a pixel art editor and a
+        sequencer. It also does animation.
       </p>
-      <p>
-        Pick a color by clicking the color button or pressing keys <code>1</code>-<code>9</code>.
-        Then start painting! You can only paint by clicking a pixel at a time, sorry :(
-      </p>
-      <p>
-        To help you, there are some keyboard modifiers. If you press <code>alt</code>/<code>option</code> while
-        painting a pixel, you will get a horizontal line. If you press <code>shift</code> while painting
-        a pixel, you will get a vertical line. Lines extend through all pixels of the same
-        color, so you can put a start and end pixel and the line will cover the area between
-        them. If you press <code>windows</code>/<code>command</code>/<code>meta</code> (depending on your OS) the line will only
-        paint every other pixel.
-      </p>
-      <p>
-        Use keys <code>a</code>-<code>y</code> to mute tracks.
-      </p>
-      <p>
-        Save works OK but I wouldn’t trust it with my life. It saves the current
-        state of your composition in the URL, just copy it from the address bar
-        to share it.
+      <details open>
+        <summary>Painting</summary>
+        <p>
+          Pick a color by clicking the color button or pressing keys
+          <code>1</code>-<code>9</code>. Then start painting! You can only paint
+          by clicking a pixel at a time, sorry :(
+        </p>
+        <p>
+          To help you, there are some keyboard modifiers. If you press
+          <code>alt</code>/<code>option</code> while painting a pixel, you will
+          get a horizontal line. If you press <code>shift</code> while painting
+          a pixel, you will get a vertical line. Lines extend through all pixels
+          of the same color, so you can put a start and end pixel and the line
+          will cover the area between them. If you hold down{' '}
+          <code>windows</code>/<code>command</code>/<code>meta</code> aswell
+          (depending on your OS) while paint a line, the line will only paint
+          every other pixel.
+        </p>
+      </details>
 
-        If you get no sound when you press PLAY, try hitting SAVE and reloading your browser.
-      </p>
-      {/*
-      <p>
-        1: Kick
-        2: Snare
-        3: Hi-hat
-        4: Synth 2
-        5: Pluck 2
-        6: Duo 1
-        7: Sampler 1
-        8: Mono 2
-        9: Fun 2
-      </p>
-      */}
+      <hr />
+
+      <details>
+        <summary>Sounds</summary>
+        <p>
+          Each color represent a specific sound. For non-drum sounds, the note
+          of the sound is controlled by the track it is painted on. The note of
+          the track is indicated in the column to the left of the track.
+        </p>
+        <p>
+          You can use keys <code>a</code>-<code>z</code> to mute tracks. The
+          mute letter of the track is indicated in the column to the right of
+          the track.
+        </p>
+      </details>
+
+      <hr />
+
+      <details>
+        <summary>Animation</summary>
+        <p>
+          If you want to, you can create multiple frames, which you can then
+          animate between. Get started by clicking the <code>+</code> button on
+          the Frames toolbar next to the color buttons.
+        </p>
+        <p>
+          For convenience, you can press <code>C</code> instead, to clone the
+          current frame into the next frame. This might help if only parts of
+          your frame is animated.
+        </p>
+        <p>
+          You could also use frames for song parts: Disable animation by
+          clicking the checkbox and switch between frames manually by using the
+          arrow buttons.
+        </p>
+        <p>
+          When using frames, be aware that the note for a specific track changes
+          per frame, to keep everything fun and random.
+        </p>
+      </details>
+
+      <hr />
+
+      <details>
+        <summary>Saving</summary>
+        <p>
+          Save works OK but I wouldn’t trust it with my life. It saves the
+          current state of your composition in the URL, just copy it from the
+          address bar to share it.
+        </p>
+        <p>
+          If you don’t get any sound when you press PLAY, try hitting SAVE and
+          reloading your browser.
+        </p>
+      </details>
+
+      <hr />
+
+      <details>
+        <summary>How do I stop this thing?!?</summary>
+        <p>
+          I’m trying to get Tone.js’ start/stop mechanism functioning reliably,
+          but I’ve not had any success so far. Until I get it to work reliably
+          and consistently, I don’t want to put a STOP button in the UI at all.
+        </p>
+        <p>
+          For now, do what I do: Save, then reload you browser. This also
+          encourages you to save often, so... you’re welcome!
+        </p>
+      </details>
+
+      <hr />
+
       <p>Version: {version}</p>
       <Show when={showVersionWarning()}>
         <p class="warning">
-          The current composition is saved with version {createdWith}. PIKSEQ
-          is running version {version}. Some sounds may not represent what the
+          The current composition is saved with version {createdWith}. PIKSEQ is
+          running version {version}. Some sounds may not represent what the
           original composer intended.
         </p>
       </Show>