+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 This changelog is automatically generated.
 
+## v0.1.4
+
+- Put help text in sections, hide all except the first [`bff5d27`](../../commit/bff5d274388dda6ce04a39c1cba2d88c83c0241b)
+
+## v0.1.3
+
+- Add button to duplicate current frame. Formatting [`bb6a554`](../../commit/bb6a5541aaedfcf2e63e2695b8d7e7d16d214bce)
+
+## v0.1.2
+
+- Add help text about animation [`7c7bebc`](../../commit/7c7bebce4015571283e0201bbecdf21f00e365d9)
+
+## v0.1.1
+
+- Set animate to false when converting old saves [`8e34308`](../../commit/8e34308f24fefc87e34dcb8a78729c44b50caad3)
+
+## v0.1.0
+
+- Fix handleTickClick function [`c724f1b`](../../commit/c724f1bb9453232ce425a0cea47331639c20a647)
+- WIP Frames [`4d57a6e`](../../commit/4d57a6edf15a745e0fb56b2b11f881d87c2e27a8)
+
+## v0.0.6
+
+- Add To do item about animation [`1761f01`](../../commit/1761f017768704c0284fd9acd4e791a552d69d81)
+- WIP: Color cycling not working [`084966e`](../../commit/084966e0dc91046c55be418cabe07fab115a9c69)
+- Show muta letter for each track. Strike out chord for muted tracks [`2f0f789`](../../commit/2f0f78968fb643c6b3a443134595d4336d8659bd)
+- Clean up Help component [`e814a9d`](../../commit/e814a9d4082ddc858791f1793a3314892cdbbc18)
+
+## v0.0.5
+
+- Try slightly larger grid and lighter background [`c55fd5a`](../../commit/c55fd5a897a1cd8211133ac8eabbff34ee582037)
+
 ## v0.0.4
 
 - Highlight note when hovering related track [`2c746ab`](../../commit/2c746ab9bd7ef2a9b21535cd591a7ff9de6430f9)